Create Rock_Paper_Scissor.py
Write a tiny game or everything else in Python to get fluent in Python. No fancy graphics needed. A tic-tac-toe on console is fine. If you already wrote Python code, send me the GitHub Link or a part of the code in the PDF with an explanation (I will Goole the code, so no copycat, please ;-)